# Ensure a state change tracker setup from inside a state change listener does not fire immediately