@bdraco 2020-07-17
Verified no substantive performance regression.  


Before
```
# hass --script benchmark state_changed_event_helper
Using event loop: _UnixSelectorEventLoop
Benchmark state_changed_event_helper done in 8.130756168160588s
Benchmark state_changed_event_helper done in 7.993554716929793s
```

After
```
# hass --script benchmark state_changed_event_helper
Using event loop: _UnixSelectorEventLoop
Benchmark state_changed_event_helper done in 7.9830495859496295s
Benchmark state_changed_event_helper done in 7.867380789946765s
```